Added a new role for provisioning VMs
Created a role provision_vm that is used to provision VMs using
virt-install. It creates a kickstart file and a script for calling
virt-install (these are based on templates) and calls the
customized virt-install script. Creating of a VM is delegated to
a hypervisor host specified in the host_vars of the virtual
machine. For example, the hypervisor for pouta-devel2 is
taito-cloud2.

Created a playbook that calls the new role (provisioning.yml).

Modified host_vars for pouta-devel1 and pouta-devel2 so that they
can be provisioned using the new role.