Change virt-install cpu flag based on OS
With CentOS 6 the correct value to give to the --cpu flag of
virt-install is "host". With CentOS 7 this value is "host-model". Added
a grouping of hypervisor hosts based on OS major version and added group
vars for both CentOS 6 and 7 hypervisors.

Also increased the wait time of virt-install from 10 minutes to 15
minutes as 10 minutes seemed to be too short in some occasions when
installing CentOS 7 as the number of packages to install is much higher
than previously with CentOS 6 and it takes longer.