Add erlang to the list of packages to install
It is useful to install erlang during build time as this makes it
possible to install rabbitmq-server using rpm, which is what the
RabbitMQ Puppet module uses by default. If erlang is not installed from
kickstart or through other means, the installation of RabbitMQ will fail
using the rpm provider. Using the rpm provider is necessary because of
the way the RabbitMQ module is written if one wants to install the
latest versions of RabbitMQ.