Make configuring a Puppet master optional
It should not be mandatory to set a Puppet master IP if the VM you are
provisioning is not going to be managed by Puppet. This was the case for
one /etc/hosts entry added in the Kickstart template. Only add the line
for the Puppet master if the puppetmaster_ip variable is defined.