Allow configurable delegation of wait_for task
Previously we always used the hypervisor of a VM as a jump host for that
VM to ensure SSH connectivity. In some cases we don't want to go through
the hypervisor but a different host instead. Create a new variable
called bastion_host for this, but use the "hyper" host as a default.