Option for extra packages and no erlang by default
Add an option that allows installation of additional packages on a host
by host basis depending on the value of a new extra_packages variable.
Also remove installation of erlang by default. It is not needed on all
provisioned VMs. If erlang is needed, it can be installed using the new
extra_packages variable.