Multiple network support for image based provision (#26)
Added support for multiple networks for image-based provisioing. Since
this changed the expected configuration data, the defaults.yml was
documented. An assertion was also added to make sure this change doesn't
silently break deployments with older data.

Added a fix to remove resolv.conf for centos, it had a 10.0.2.3 dns
server there by default which did not get overwritten.